@@ -1,6 +1,6 @@
 {
   "name": "@jfparent/ng-mat-search-bar",
-  "version": "13.0.2",
+  "version": "14.0.0",
   "license": "MIT",
   "scripts": {
     "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points --tsconfig tsconfing.base.json",
@@ -17,19 +17,19 @@
     "url": "https://github.com/jifparent/ng-mat-search-bar.git"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-devkit/core": "^13.3.11",
-    "@angular/animations": "^13.4.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/cli": "^13.3.11",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-devkit/core": "^14.2.13",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/cli": "^14.2.13",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
     "@types/jasmine": "^3.10.14",
     "@types/jasminewd2": "^2.0.11",
     "@types/node": "^12.20.55",
@@ -43,7 +43,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "^4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "rxjs": "^6.6.7",
     "ts-node": "~8.3.0",
@@ -51,12 +51,7 @@
     "typescript": "~4.6.4",
     "zone.js": "^0.11.8"
   },
-  "peerDependencies": {
-    "@angular/animations": "^13.0.0",
-    "@angular/core": "^13.0.0",
-    "@angular/forms": "^13.0.0",
-    "@angular/material": "^13.0.0"
-  },
+  "peerDependencies": {},
   "ngPackage": {
     "lib": {
       "entryFile": "public_api.ts"
